+++ portlet-common/src/main/java/de/seitenbau/govdata/constants/GovDataConfigParam.java
@@ -1,0 +1,20 @@
+package de.seitenbau.govdata.constants;
+
+public abstract class GovDataConfigParam
+{
+  /** Der Parameter "fusekiUrl". */
+  public static final String FUSEKI_URL = "fusekiUrl";
+
+  /** Der Parameter "fusekiSparqlEndpoint". */
+  public static final String FUSEKI_SPARQL_ENDPOINT = "fusekiSparqlEndpoint";
+
+  /** Der Parameter "cKANurl". */
+  public static final String CKAN_URL = "cKANurl";
+
+  /** Der Parameter "cKANurlFriendly". */
+  public static final String CKAN_URL_FRIENDLY = "cKANurlFriendly";
+
+  /** Der Parameter "fusekiDatastoreName". */
+  public static final String FUSEKI_DATASTORE_NAME = "fusekiDatastoreName";
+
+}